Trait query_tests_setup::ConnectorTagInterface

pub trait ConnectorTagInterface {
    fn datamodel_provider(&self) -> &'static str;

    fn datamodel_renderer(&self) -> Box<dyn DatamodelRenderer>;

    fn connection_string(&self, test_database: &str, is_ci: bool) -> String;

    fn capabilities(&self) -> &[ConnectorCapability];

    fn as_parse_pair(&self) -> (String, Option<String>);

    fn is_versioned(&self) -> bool;

    fn referential_integrity(&self) -> &'static str { ... }
}

Required methods

fn datamodel_provider(&self) -> &'static str

The name of the datamodel provider for this connector. Must match valid datamodel provider strings.

fn datamodel_renderer(&self) -> Box<dyn DatamodelRenderer>

Returns the renderer to be used for templating the datamodel (the models portion).

fn connection_string(&self, test_database: &str, is_ci: bool) -> String

The connection string to use to connect to the test database and version.

test_database is the database to connect to, which is an implementation detail of the implementing connector, like a file or a schema.

is_ci signals whether or not the test run is done on CI or not. May be important if local test run connection strings and CI connection strings differ because of networking.

fn capabilities(&self) -> &[ConnectorCapability]

Capabilities of the implementing connector.

fn as_parse_pair(&self) -> (String, Option<String>)

Serialization of the connector. Expected to return (tag_name, version). Todo: Think of something better.

fn is_versioned(&self) -> bool

Must return true if the connector family is versioned (e.g. Postgres9, Postgres10, …), false otherwise.

Provided methods

fn referential_integrity(&self) -> &'static str

Defines where relational constraints are handled:

“prisma” is handled in the Query Engine core

“foreignKeys” lets the database handle them