Crate query_tests_setup

Re-exports

pub use query_core;

Structs

CONFIG

Test configuration, loaded once at runtime.

DatamodelWithParams

Directive

A field directive, e.g. @map("_id").

ENGINE_PROTOCOL

Engine protocol used to run tests. Either ‘graphql’ or ‘json’.

ENV_LOG_LEVEL

The log level from the environment.

IdFragment

ID field definition, e.g. #id(id, Int, @id @test.SmallInt)

JsonRequest

JsonResponse

M2mFragment

M2m field definition, e.g. #m2m(posts, Post[], id, String, "name")

MongoDbConnectorTag

MongoDbSchemaRenderer

MultiQuery

MySqlConnectorTag

PostgresConnectorTag

PrintWriter

This is a temporary implementation detail for tracing logs in tests. Instead of going through std::io::stderr, it goes through the specific local stderr handle used by eprintln and dbg, allowing logs to appear in specific test outputs for readability.

QueryParams

QueryParams enables parsing the generated id(s) of mutations sent to the Query Engine so that it can be reused in subsequent queries

QueryResult

Runner

Direct engine runner.

SqlDatamodelRenderer

SqlServerConnectorTag

SqliteConnectorTag

TemplatingError

TestConfig

The central test configuration.

TestLogCapture

VitessConnectorTag

Enums

ConnectorTag

ConnectorVersion

DatamodelFragment

A datamodel fragment is the parsed version of a #<ident>(...) template string.

FragmentArgument

Wrapper for general argument parsing.

GraphqlBody

Identifier

MongoDbVersion

MySqlVersion

PostgresVersion

QueryParamsWhere

QueryParamsWhereMany

RelationField

RelationReference

SqlServerVersion

TemplatingErrorKind

TestError

VitessVersion

Constants

FULL_ID_OPTIONS

SIMPLE_ID_OPTIONS

Traits

AsyncFn

Taken from Reddit. Enables taking an async function pointer which takes references as param https://www.reddit.com/r/rust/comments/jvqorj/hrtb_with_async_functions/

ConnectorTagInterface

ConnectorTestFn

DatamodelRenderer

The main trait a datamodel renderer for a connector has to implement.

VecUtilities

Functions

child_references

common_child_references

common_parent_references

full_child_references

full_parent_references

make_json_custom_type

parent_references

parse

Main entry point into the template parsing. Parses a schema fragment of the form #<fragment_ident>...<eol>.

parse_compound_id

Parses the JSON result of mutation sent to the Query Engine in order to extract the generated compound ids. Returns a string that’s already formatted to be included in another query. eg: { “id_1_id_2”: { id_1: “my_fancy_id_1”, id_2: “my_fancy_id_2” } }

parse_id

Parses the JSON result of mutation sent to the Query Engine in order to extract the generated id. Returns a string that’s already formatted to be included in another query. eg: { “id”: “my_fancy_id” }

parse_many_compound_ids

Performs the same extraction as parse_compound_id but for an array

parse_many_ids

Performs the same extraction as parse_id but for an array

render_test_datamodel

Render the complete datamodel with all bells and whistles.

run_connector_test

run_connector_test_impl

run_relation_link_test

run_with_tokio

Helper method to allow a sync shell function to run the async test blocks.

schema_with_relation

setup_metrics

setup_project

Setup of everything as defined in the passed datamodel.

simple_child_references

simple_parent_references

teardown_project

Teardown of a test setup.

test_tracing_subscriber

walk_json

Type Definitions

DatamodelsAndCapabilities

LogEmit

TemplatingResult

TestResult

TxResult